% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_theme.R
\name{add_theme}
\alias{add_theme}
\title{Add theme to ggplot}
\usage{
add_theme(plot, SIZE = 5)
}
\arguments{
\item{plot}{plot}

\item{SIZE}{size}
}
\description{
Add theme to ggplot
}
\examples{
plot \%>\% add_theme()
}
\keyword{add}
\keyword{plot}
\keyword{theme}
\keyword{to}
