% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{bar plot ggplot2}
\usage{
bar_plot(plot, theme = TRUE)
}
\arguments{
\item{plot}{plot where mean is dependent variable}

\item{theme}{add theme?}
}
\description{
bar plot ggplot2
}
\examples{
plot \%>\% bar_plot
}
\keyword{a}
\keyword{bar}
\keyword{make}
\keyword{plot}
