% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{A helper function for onefactor and twofactor}
\usage{
barplot(bar.arg, limits, breaks)
}
\arguments{
\item{bar.arg}{box.arg}

\item{limits}{limits}

\item{breaks}{breaks}
}
\description{
Use onefactor and twofactor instead of this
}
\examples{
barplot(graph, limits, breaks)
}
\keyword{boxplot}
