% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{A helper function for onefactor and twofactor}
\usage{
boxplot(box.arg, y)
}
\arguments{
\item{box.arg}{box.arg}

\item{y}{dependent variable}
}
\description{
Use onefactor and twofactor instead of this
}
\examples{
boxplot(graph, limits, breaks)
}
\keyword{boxplot}
