% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohend.R
\name{cohend}
\alias{cohend}
\title{A function that will give the confidence interval around Cohen's d.
It requires an estimate of d, n1, n2, and whether or not you want the upper/lower bound}
\usage{
cohend(d, n1, n2, alpha = 0.95)
}
\arguments{
\item{d}{estimate of Cohen's d}

\item{n1}{sample size of cell 1}

\item{n2}{sample size of cell 2}

\item{alpha}{alpha level}
}
\description{
A function that will give the confidence interval around Cohen's d.
It requires an estimate of d, n1, n2, and whether or not you want the upper/lower bound
}
\examples{
cohend(d.est, n1, n2)
}
\keyword{effect}
\keyword{size}
