% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographics}
\alias{demographics}
\title{A function that gives the demographics of a dataset}
\usage{
demographics(dataset)
}
\arguments{
\item{dataset}{A dataframe with subject, gender, age as variables.}
}
\description{
This function will give you a summary of the demographics of a dataset.
It requires the following columns: dataset$subject, dataset$gender, dataset$age
dataset$gender can have values 0,1,2,3 or na, male, female, other
}
\examples{
demographics()
}
\keyword{demographics}

