% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualtrics.R
\name{get_qualtrics}
\alias{get_qualtrics}
\title{A function get Qualtrics data}
\usage{
get_qualtrics(surveyid, apitoken, rooturl, force = FALSE, save = FALSE)
}
\arguments{
\item{surveyid}{survey id}

\item{apitoken}{api token}

\item{rooturl}{root url}

\item{force}{logical: re-download?}

\item{save}{logical: save?}
}
\description{
needs survey id, api token, root url, and two optional parameters allow to force to re-download data & to save the data
}
\examples{
get_qualtrics("qmlskdfj","mkjqmdlsfkjsd","url",force = FALSE, save = TRUE)
}
\keyword{Qualtrics}
\keyword{get}
