% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.barplot.R
\name{ggplot2.barplot}
\alias{ggplot2.barplot}
\title{A helper function for onefactor and twofactor}
\usage{
ggplot2.barplot(bar.arg, lwr, upr)
}
\arguments{
\item{bar.arg}{box.arg}

\item{lwr}{lwr end of ci}

\item{upr}{upr end of ci}
}
\description{
Use onefactor and twofactor instead of this
}
\examples{
barplot(graph, limits, breaks)
}
\keyword{boxplot}
