% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.experiment.R
\name{graph.experiment}
\alias{graph.experiment}
\title{Boxplots and barplots for two X variables}
\usage{
graph.experiment(y, x, dataset, graph.type = "bar", ws = 0, SIZE = 5)
}
\arguments{
\item{y}{dependent variable}

\item{x}{vector with names of independent variables}

\item{graph.type}{boxplot or barplot}
}
\description{
Boxplots and barplots for two X variables
}
\examples{
graph.experiment("control", c("info","imagination"), study3, graph.type = "box")
}
\keyword{barplot}
\keyword{boxplot}
