% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.fake.results.R
\name{graph.fake.results}
\alias{graph.fake.results}
\title{Generate graph based on means (for class)}
\usage{
graph.fake.results(a, b, c, d, scale, contrast, context, dv, SIZE = 5,
  addlabels = 0)
}
\arguments{
\item{a}{mean}

\item{b}{mean}

\item{c}{mean}

\item{d}{mean}

\item{scale}{scale, start from 0}

\item{contrast}{first factor}

\item{context}{second factor}

\item{dv}{label}

\item{addlabels}{add labels}
}
\description{
Generate graph based on means (for class)
}
\examples{
graph.fake.results(a = 2, b = 5, c = 7, d = 2, scale = c(0,7), contrast = c("gender","male","female"), context = c("social_influence","absent","present"), "choice")
}
\keyword{barplot}
