% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationsummary.R
\name{mediation.summary}
\alias{mediation.summary}
\title{A function that summarizes the output of a bootstrap procedure in a mediation analysis}
\usage{
mediation.summary(output)
}
\arguments{
\item{output}{a dataset generated by a bootstrap procedure}
}
\description{
This function summarizes the columns in an output dataset generated by a bootstrap procedure.
}
\examples{
mediation.summary()
}
\keyword{bootstrap}
\keyword{mediation}
