% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationanalysis.R
\name{mediationanalysis}
\alias{mediationanalysis}
\title{Simple mediation analysis}
\usage{
mediationanalysis(y, mediator, x, effects, data, simulate = 0,
  numberofsimulations = 5000, seed = 123, alpha = 0.95)
}
\arguments{
\item{y}{dependent variable}

\item{mediator}{mediator}

\item{x}{independent variable(s)}

\item{effects}{which effect are we testing}

\item{data}{dataset}

\item{simulate}{calculate CIs via bootstrapping?}

\item{numberofsimulations}{number of bootstraps}

\item{seed}{set seed to get the same results}

\item{alpha}{significance level}
}
\description{
Simple mediation analysis
}
\examples{
graph.experiment("regard", "impression", "cc", "ccgucci")
}
\keyword{mediation}
