% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefactor.R
\name{onefactor}
\alias{onefactor}
\title{Boxplots and barplots for one X variable}
\usage{
onefactor(y, x1, graph.type = "bar")
}
\arguments{
\item{y}{dependent variable}

\item{x1}{independent variable}

\item{graph.type}{boxplot or barplot}
}
\description{
Boxplots and barplots for one X variable
}
\examples{
onefactor(data$y, data$x, graph.type = "box")
}
\keyword{barplot}
\keyword{boxplot}
