% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{A function to read qualtrics CSV files}
\usage{
read_qualtrics(x)
}
\arguments{
\item{x}{csv filename}
}
\description{
You need to pass on the name of the csv file
}
\examples{
read_qualtrics("data.csv")
}
\keyword{Qualtrics}
\keyword{read}
