% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics_csv.R
\name{read_qualtrics_csv}
\alias{read_qualtrics_csv}
\title{Read Qualtrics data CSV files}
\usage{
read_qualtrics_csv(filename)
}
\arguments{
\item{filename}{filename (include .csv extension)}
}
\description{
Read Qualtrics data CSV files
}
\examples{
read_qualtrics_csv("dataset.csv")
}
\keyword{taskmaster}
