% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{A function that removes duplicates from a dataset}
\usage{
remove_duplicates(dataset)
}
\arguments{
\item{dataset}{A dataframe with experiment (optional), subject, MID as variables.}
}
\description{
It requires the following columns: dataset$experiment (optional), dataset$subject, dataset$MID
}
\details{
data$experiment =1 for the experiment that happened first chronologically, = 2 for the experiment that happened second, ...

dataset$subject = 1 for the subject that participated first chronologically, = 2 for the subject that participated second, ...

dataset$MID refers to mturk (or other) IDs
}
\examples{
dataset <- remove_duplicates(dataset)
}
\keyword{duplicates}

