% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{savegraph}
\alias{savegraph}
\title{Save graph}
\usage{
savegraph(filename, h = 1, w = 1)
}
\arguments{
\item{filename}{filename}

\item{h}{height}

\item{w}{width}
}
\description{
Save graph
}
\keyword{a}
\keyword{are}
\keyword{dimensions}
\keyword{for}
\keyword{graph}
\keyword{in}
\keyword{perfect}
\keyword{powerpoint}
\keyword{save}
\keyword{that}
