% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spssanova.R
\name{spssanova}
\alias{spssanova}
\title{A function that generates the same ANOVA output as SPSS does.
This is a BETA version.
This will set your contrast coding to treatment
This will give a different output when first assigning the linear model to an object and then applying the function (e.g., z <- lm(y~x) ; spssanova(z))
vs when not assigning the linear model to an object (e.g., spssanova(lm(y~x)))
I don't know why this happens, but the second method gives the same output as SPSS}
\usage{
spssanova(linearmodelfit)
}
\arguments{
\item{linearmodelfit}{A linear model}
}
\description{
A function that generates the same ANOVA output as SPSS does.
This is a BETA version.
This will set your contrast coding to treatment
This will give a different output when first assigning the linear model to an object and then applying the function (e.g., z <- lm(y~x) ; spssanova(z))
vs when not assigning the linear model to an object (e.g., spssanova(lm(y~x)))
I don't know why this happens, but the second method gives the same output as SPSS
}
\examples{
spssanova(lm(y ~ x, data))
}
\keyword{anova}
\keyword{spss}
