% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_for_graph.R
\name{summarize_for_graph}
\alias{summarize_for_graph}
\title{Summarize data set}
\usage{
summarize_for_graph(data, yvar)
}
\arguments{
\item{data}{data}

\item{yvar}{variable}
}
\description{
Summarize data set
}
\examples{
data \%>\% summarize_for_graph(measure)
}
\keyword{summarize}
