% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{A function that "Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95 percent)"}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{measurevar}{the name of a column that contains the variable to be summarized}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{conf.interval}{the percent range of the confidence interval (default is 95 percent)}
}
\description{
Not my function. 
Got it from: http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/
Requires library(plyr)
}
\examples{
tgc <- summarySE(tg, measurevar="len", groupvars=c("supp","dose"))
}
\keyword{ggplot2}
\keyword{summary}

