% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twofactors.R
\name{twofactors}
\alias{twofactors}
\title{Boxplots and barplots for two X variables}
\usage{
twofactors(y, x1, x2, graph.type = "bar")
}
\arguments{
\item{y}{dependent variable}

\item{x1}{independent variable}

\item{x2}{another independent variable}

\item{graph.type}{boxplot or barplot}
}
\description{
Boxplots and barplots for two X variables
}
\examples{
onefactor(data$y, data$x, graph.type = "box")
}
\keyword{barplot}
\keyword{boxplot}
