% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{wide2long}
\alias{wide2long}
\title{A function to convert datasets from wide to long format}
\usage{
wide2long(x, measure.vars)
}
\arguments{
\item{x}{dataset}

\item{measure.vars}{wide variables that will become a long variable}
}
\description{
You need to pass on the dataset and a vector with 'measure.vars', i.e., the wide variables that will become a long variable
}
\examples{
wide2long(data, c("measure1","measure2"))
}
\keyword{format}
\keyword{long}
\keyword{wide}
