% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writecontrast.R
\name{writecontrast}
\alias{writecontrast}
\title{A function for calculating (complicated) contrasts}
\usage{
writecontrast(n1, n2, nrofterms, aovmodel, K, rmd = FALSE, list = FALSE)
}
\arguments{
\item{n1}{number of observations in cell 1}

\item{n2}{number of observations in cell 2}

\item{nrofterms}{how many terms are involved?}

\item{aovmodel}{An anova model}

\item{K}{contrast}

\item{Y}{dependent variable}
}
\description{
This function will calculate a contrast and summarize it in publication-ready output.
This is for complicated contrasts. For simple contrasts (simply comparing two cell means), use writesimplecontrast
Set rmd to true to get Rmarkdown output.
Set list to true to return a list object.
RMD trumps list.
}
\examples{
writecontrast()
}
\keyword{contrast}
