% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writesimplecontrast.R
\name{writesimplecontrast}
\alias{writesimplecontrast}
\title{A function for comparing two cell means}
\usage{
writesimplecontrast(x, y, aovmodel, K, rmd = FALSE, list = FALSE)
}
\arguments{
\item{x}{which experimental cell 1}

\item{y}{which experimental cell 2}

\item{aovmodel}{An anova model}

\item{K}{contrast}
}
\description{
This function will calculate a contrast and summarize it in publication-ready output.
This is for simple contrasts (simply comaring two cell means). For more complicated contrasts, use writecontrast.
You only need to pass on two numbers, x and y. These numbers refer to the experimental cells. They denote the subscripts of the Ns, the means, and the sds table.
Ns should contain the sample sizes, means should contain the means, sds should contain the standard deviations
Set rmd to true to get Rmarkdown output.
Set list to true to return a list object.
RMD trumps list.
}
\examples{
writecontrast()
}
\keyword{contrast}
