% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationanalysis.R
\name{mediationanalysis}
\alias{mediationanalysis}
\title{Simple mediation analysis}
\usage{
mediationanalysis(
  y,
  mediator,
  x_variables,
  data,
  simulate = FALSE,
  numberofsimulations = 5000,
  seed = 123,
  alpha = 0.95
)
}
\arguments{
\item{y}{dependent variable}

\item{mediator}{mediator}

\item{x_variables}{independent variable(s)}

\item{data}{dataset}

\item{simulate}{calculate CIs via bootstrapping?}

\item{numberofsimulations}{number of bootstraps}

\item{seed}{set seed to get the same results}

\item{alpha}{significance level}
}
\description{
Simple mediation analysis
}
\examples{
mediationanalysis(y = "tour", mediator = "control", x_variables = c("imagery_appeal","x2"), simulate = TRUE, data = data)
}
\keyword{mediation}
