% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spssanova.R
\name{spssanova}
\alias{spssanova}
\title{A function that generates the same ANOVA output as SPSS does, i.e., Type III Anova}
\usage{
spssanova(linearmodel)
}
\arguments{
\item{linearmodel}{A linear model.}
}
\description{
A function that generates the same ANOVA output as SPSS does, i.e., Type III Anova
}
\examples{
spssanova(lm(y ~ x1 * x2, data=data))
}
\keyword{anova}
\keyword{spss}
