% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spssanova.R
\name{spssanova}
\alias{spssanova}
\title{A function that generates the same ANOVA output as SPSS does, i.e., Type III Anova}
\usage{
spssanova(formula, DATA)
}
\arguments{
\item{formula}{A formula.}

\item{DATA}{data}
}
\description{
A function that generates the same ANOVA output as SPSS does, i.e., Type III Anova
}
\examples{
spssanova(formula = "p1 ~ conditie", DATA = md)
}
\keyword{anova}
\keyword{spss}
